+++ HVLib/HTTP/HVHttp.m
@@ -82,14 +82,15 @@ static NSString* const c_header_contentRange = @"Content-Range";
 
 -(void)setContentLength:(NSUInteger)length
 {
-    NSString* value = [NSString stringWithFormat: @"%d", length];
+    NSString* value = [NSString stringWithFormat: @"%lu", (unsigned long)length];
     [self setValue:value forHTTPHeaderField:c_header_contentLength];
 }
 
 -(void) setContentRangeStart:(NSUInteger) start end:(NSUInteger) end
 {
-    NSString* value = [NSString stringWithFormat:@"bytes %d-%d/*", start, end];
+    NSString* value = [NSString stringWithFormat:@"bytes %lu-%lu/*", (unsigned long)start, (unsigned long)end];
     [self setValue:value forHTTPHeaderField:c_header_contentRange];
+    
 }
 
 -(void)setContentType:(NSString *)type
